% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{chenAr}
\alias{chenAr}
\title{Chen Areal Atlas}
\usage{
chenAr()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Brain atlas for the genetic topography of cortical areal morphology
with 12 regions per hemisphere.
Contains 2D polygon geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}}.
}
\examples{
chenAr()
}
\references{
Chen et al. (2013) PNAS, 110 (42) 17089-17094;
\doi{10.1073/pnas.1308091110}
}
\seealso{
Other ggseg_atlases: 
\code{\link{chenTh}()}
}
\concept{ggseg_atlases}
