% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegChen-package.R
\docType{package}
\name{ggsegChen-package}
\alias{ggsegChen}
\alias{ggsegChen-package}
\title{ggsegChen: Chen Atlas for the 'ggseg' Ecosystem}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Chen cortical atlases for the 'ggseg' ecosystem. Provides unified 'ggseg_atlas' objects with 2D polygon geometry for use with 'ggseg'. Includes both thickness (chenTh) and areal (chenAr) parcellations from Chen et al. (2013).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggsegChen}
  \item Report bugs at \url{https://github.com/ggsegverse/ggsegChen/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Didac Vidal-Pineiro \email{d.v.pineiro@psykologi.uio.no} (\href{https://orcid.org/0000-0001-9997-9156}{ORCID})
}

}
\keyword{internal}
