% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chen.R
\name{chenAr}
\alias{chenAr}
\title{Chen Area Cortical Atlas}
\usage{
chenAr()
}
\value{
A [ggseg.formats::ggseg_atlas] object (cortical).
}
\description{
Genetic topography of brain area morphology with 12 cortical regions
per hemisphere based on Chen et al. (2013).
}
\examples{
chenAr()
}
\references{
Chen C-H, et al. (2013).
  Genetic topography of brain morphology.
  *PNAS*, 110(42), 17089-17094.
  \doi{10.1073/pnas.1308091110}
}
\seealso{
Other ggseg_atlases: 
\code{\link{chenTh}()}

Other cortical_atlases: 
\code{\link{chenTh}()}
}
\concept{cortical_atlases}
\concept{ggseg_atlases}
