% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chen.R
\name{chenTh}
\alias{chenTh}
\title{Chen Thickness Cortical Atlas}
\usage{
chenTh()
}
\value{
A [ggseg.formats::ggseg_atlas] object (cortical).
}
\description{
Genetic topography of brain thickness morphology with 12 cortical
regions per hemisphere based on Chen et al. (2013).
}
\examples{
chenTh()
}
\references{
Chen C-H, et al. (2013).
  Genetic topography of brain morphology.
  *PNAS*, 110(42), 17089-17094.
  \doi{10.1073/pnas.1308091110}
}
\seealso{
Other ggseg_atlases: 
\code{\link{chenAr}()}

Other cortical_atlases: 
\code{\link{chenAr}()}
}
\concept{cortical_atlases}
\concept{ggseg_atlases}
